[clean] move geolens package files into a folder
