+++ 2_autolens_rms.py
@@ -22,7 +22,7 @@ import yaml
 from tqdm import tqdm
 
 from deeplens.geolens import GeoLens
-from deeplens.geolens_utils import create_lens
+from deeplens.geolens_pkg.utils import create_lens
 from deeplens.optics.basics import DEPTH, EPSILON, WAVE_RGB
 from deeplens.utils import create_video_from_images, set_logger, set_seed
