+++ deeplens/__init__.py
@@ -4,6 +4,9 @@ from .optics import *
 # network
 from .network import *
 
+# geolens
+from .geolens_pkg import *
+
 # utilities
 from .utils import *
